requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain-openai
jinja2
python-multipart